+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
 Release notes
 ==============
 
+### Unreleased
+
+* Fixed missing primary key in temp table for `mergeBatch` table scripts
+
 ### 2.6.0 (2022-12-10)
 
 * Added batch insert/update/merge table scripts (`insertBatch`, `updateBatch`, and `mergeBatch`)